# docs/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: AgentCrew Sidecar API
  version: 0.5.0-draft
  description: |
    单用户本地 sidecar。M0 操作带完整请求/响应/错误 schema；
    M2 操作定义当前身份、资源、授权、审计和诊断边界；实施归属见 x-domain-card。
    演示身份通过服务端签发的 X-AgentCrew-Identity 标识固定于每个请求。
    p3/p5 操作 schema 随对应切片细化（当前为方向性描述）。
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port: { default: "8710" }
security:
  - bearerToken: []
  - bearerToken: []
    demoIdentity: []
...
components:
  securitySchemes:
    bearerToken: { type: http, scheme: bearer }
    demoIdentity: { type: apiKey, in: header, name: X-AgentCrew-Identity, description: 服务端签发的演示身份标识，Bearer仍然必需 }
  schemas:
...
    SkillCreateRequest:
      type: object
      additionalProperties: false
      required: [change_id, expected_revision, workspace_id, agent_id, name, description, text, basis]
      properties:
        change_id: { type: string, minLength: 1, maxLength: 128 }
        expected_revision: { type: integer, const: 0 }
        workspace_id: { type: string, minLength: 1 }
        agent_id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, minLength: 1, maxLength: 60 }
        text: { type: string, minLength: 1 }
        basis: { type: string, minLength: 1 }
        conversation_id: { type: [string, "null"] }
        files: { type: object, additionalProperties: { type: string } }
    GovernanceChangeRequest:
      type: object
      required: [change_id, expected_revision]
      properties:
        change_id: { type: string, minLength: 1, maxLength: 128 }
        expected_revision: { type: integer, minimum: 0 }
...
    GovernanceResource:
      type: object
      required: [id, name, status, revision, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        org_id: { type: string }
        workspace_id: { type: string }
        data_dir: { type: string, description: 服务分配的工作区目录 }
        status: { type: string, enum: [active, disabled, archived] }
        revision: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        spec: { $ref: "#/components/schemas/AgentSpec" }
        description: { type: string, maxLength: 60 }
        source: { type: string, enum: [user, agent, system] }
        current_version_id: { type: [string, "null"] }
        type: { type: string, enum: [http, mcp] }
        config: { $ref: "#/components/schemas/ConnectorConfig" }
        credential_configured: { type: boolean }
    WorkspaceRequest:
      allOf:
        - { $ref: "#/components/schemas/GovernanceChangeRequest" }
        - type: object
          required: [name]
          properties:
            name: { type: string, minLength: 1, maxLength: 80 }
            status: { type: string, enum: [active, disabled, archived] }
    AgentSpec:
      type: object
      additionalProperties: false
      required: [position, model_slot, skill_ids, connector_ids]
      properties:
        position: { type: string, minLength: 1, maxLength: 2000 }
        model_slot: { type: string, enum: [main, aux] }
        skill_ids: { type: array, uniqueItems: true, items: { type: string } }
        connector_ids: { type: array, uniqueItems: true, items: { type: string } }
    AgentRequest:
      allOf:
        - { $ref: "#/components/schemas/GovernanceChangeRequest" }
        - type: object
          required: [workspace_id, name, spec]
          properties:
            workspace_id: { type: string }
            name: { type: string, minLength: 1, maxLength: 80 }
            status: { type: string, enum: [active, disabled, archived] }
            spec: { $ref: "#/components/schemas/AgentSpec" }
    GovernanceResourcePatch:
      allOf:
        - { $ref: "#/components/schemas/GovernanceChangeRequest" }
        - type: object
          properties:
            name: { type: string, minLength: 1, maxLength: 80 }
            status: { type: string, enum: [active, disabled, archived] }
            spec: { $ref: "#/components/schemas/AgentSpec" }
            description: { type: string, maxLength: 60 }
            config: { $ref: "#/components/schemas/ConnectorConfig" }
            credential: { type: [string, "null"], writeOnly: true, description: 服务端绑定认证信息；null清除，缺省不变 }
...
    GovernanceSkillCreate:
      allOf:
        - { $ref: "#/components/schemas/SkillCreateRequest" }
...
    ConnectorToolPolicy:
      type: object
      additionalProperties: false
      required: [read_only, destructive, risk_level, needs_approval]
      properties:
        read_only: { type: boolean }
        destructive: { type: boolean }
        risk_level: { type: string, enum: [low, medium, high] }
        needs_approval: { type: boolean }
        timeout_ms: { type: integer, minimum: 1, maximum: 300000, default: 60000 }
    ConnectorConfig:
      type: object
      additionalProperties: false
      required: [allowed_hosts, allowed_ports, allow_loopback]
      properties:
        url: { type: string, format: uri }
        allowed_hosts: { type: array, uniqueItems: true, items: { type: string, minLength: 1 } }
        allowed_ports: { type: array, uniqueItems: true, items: { type: integer, minimum: 1, maximum: 65535 } }
        allow_loopback: { type: boolean, default: false }
        transport: { type: string, enum: [stdio, streamable_http] }
        command: { type: string, description: stdio 固定管理员批准的可执行路径 }
        args: { type: array, items: { type: string } }
        startup_files: { type: array, maxItems: 20, items: { type: string }, description: stdio明确登记的绝对代码文件；规范化目标与SHA绑定配置修订，变更需要提交新修订 }
        credential_header: { type: string, default: Authorization }
        tool_policies: { type: object, additionalProperties: { $ref: "#/components/schemas/ConnectorToolPolicy" } }
        idempotent_endpoints:
          type: array
          items:
            type: object
            required: [method, path, guarantee]
            properties:
              method: { type: string, enum: [POST, PUT, PATCH, DELETE] }
              path: { type: string }
              guarantee: { type: string, minLength: 1, description: 上游真实去重承诺及验证依据 }
...
  parameters:
...
    GovernanceWorkspace: { name: workspace_id, in: query, schema: { type: string, minLength: 1 } }
    GovernanceLimit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    GovernanceAfter: { name: after, in: query, schema: { type: string, minLength: 1 }, description: 排他游标绑定有效身份、范围、过滤及稳定排序 }
...
  responses:
    GovernanceResourceEnvelope:
      description: 真实资源及修订，凭据只返回configured状态
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: "#/components/schemas/GovernanceResource" }
    GovernanceResourcePage:
      description: 当前身份可见资源的稳定分页
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data:
                type: object
                required: [items, next_after]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/GovernanceResource" } }
                  next_after: { type: [string, "null"] }
...
paths:
...
  /api/organization:
    get:
      operationId: readOrganization
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      responses:
        <<: *memory_errors
        "200": { $ref: "#/components/responses/GovernanceResourceEnvelope" }
    patch:
      operationId: editOrganization
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-03
      summary: owner修改组织，禁用保留历史并禁止后续业务
      requestBody: &resource_patch_body
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GovernanceResourcePatch" }
      responses:
        <<: *memory_errors
        "200": { $ref: "#/components/responses/GovernanceResourceEnvelope" }
...
  /api/workspaces:
    get:
      operationId: listWorkspaces
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      parameters: *governance_page_parameters
      responses: &resource_page_responses
        <<: *memory_errors
        "200": { $ref: "#/components/responses/GovernanceResourcePage" }
    post:
      operationId: createWorkspace
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceRequest" }
      responses: &resource_create_responses
        <<: *memory_errors
        "200": { $ref: "#/components/responses/GovernanceResourceEnvelope" }
  /api/workspaces/{id}:
    parameters: *governance_id_parameters
    get:
      operationId: readWorkspace
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      responses: *resource_create_responses
    patch:
      operationId: editWorkspace
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      requestBody: *resource_patch_body
      responses: *resource_create_responses
    delete:
      operationId: disableWorkspace
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      requestBody: &governance_change_body
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GovernanceChangeRequest" }
      responses: *resource_create_responses
  /api/agents:
    get:
      operationId: listAgents
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-05
      parameters: &workspace_page_parameters
        - { $ref: "#/components/parameters/GovernanceWorkspace" }
        - { $ref: "#/components/parameters/GovernanceLimit" }
        - { $ref: "#/components/parameters/GovernanceAfter" }
      responses: *resource_page_responses
    post:
      operationId: createAgent
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentRequest" }
      responses: *resource_create_responses
  /api/agents/{id}:
    parameters: *governance_id_parameters
    get:
      operationId: readAgent
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-05
      responses: *resource_create_responses
    patch:
      operationId: editAgent
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-02
      requestBody: *resource_patch_body
      responses: *resource_create_responses
    delete:
      operationId: disableAgent
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-05
      requestBody: *governance_change_body
      responses: *resource_create_responses
  /api/skills:
    get:
      operationId: listGovernanceSkills
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-04
      parameters: *workspace_page_parameters
      responses: *resource_page_responses
    post:
      operationId: createGovernanceSkill
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-04
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GovernanceSkillCreate" }
      responses: *resource_create_responses
  /api/skills/{id}:
    parameters: *governance_id_parameters
    get:
      operationId: readGovernanceSkill
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-04
      responses: *resource_create_responses
    patch:
      operationId: editGovernanceSkill
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-04
      requestBody: *resource_patch_body
      responses: *resource_create_responses
    delete:
      operationId: archiveGovernanceSkill
      tags: [governance]
      x-implementation-card: M2-11
      x-domain-card: M2-04
      requestBody: *governance_change_body
      responses: *resource_create_responses
...